added new import validation structure
